+++ includes/BuiltInPalettes/BrickColors.lua
@@ -8,9 +8,12 @@ for i = 1, 1032 do
 
     -- BrickColors that don't exist default to #194
     if ((brickColor.Number ~= 194) or (i == 194)) then
+        local color = brickColor.Color
+
         table.insert(brickColorsPalette.colors, {
             name = brickColor.Name,
-            color = brickColor.Color
+
+            color = { color.R * 255, color.G * 255, color.B * 255 }
         })
     end
 end